Create a new refactor file, create a module with a name of Refactor, move the main methods inside the file
